server/requirements.txt
fastapi
uvicorn
python-dotenv
openai
databases
asyncpg
sqlalchemy
pydantic
python-jose
dotenv
google-generativeai